Restructure Calculate
