+++ src/components/Calculate/CalculatePage.js
@@ -1,9 +1,9 @@
 import React, { Component } from "react";
 import Button from "@material-ui/core/Button";
-import Expense from "./Expense";
-import InputComponent from "./InputComponent";
+import Expense from "../Expense";
+import InputComponent from "../InputComponent";
 import Snackbar from "@material-ui/core/Snackbar";
-import { addParticipant, collectData } from "../actions/index";
+import { addParticipant, collectData } from "../../actions/index";
 import { connect } from "react-redux";
 
 class CalculatePage extends Component {
